Draft of Product import & Add ModelCSV, Model base classes
